pyproject.toml
[project]
name = "scraping-tabnet"
version = "0.1.0"
description = "DATASUS TabNet scraper para Produção Ambulatorial do SUS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]

[project.scripts]
scrape = "scraping_tabnet.scraper:main"
load-db = "scraping_tabnet.load_db:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/scraping_tabnet"]
